Desigining for resilience example
Add a css class to button. This css class is set to display:
none (as a default).
Add a connect method to Stimulus controller that checks is
the 'copy' command is supported, and if true adds another css
class to button. This css class sets display: initial.
This means the button is not shown if the related
functionality is not available in the browser.